Moves v1 move actions into v1 module
